refactor: update로 변경
