--- application-infrastructure/src/main/kotlin/hs/kr/equus/application/global/kafka/KafkaTopics.kt
@@ -1,7 +1,0 @@
-package hs.kr.equus.application.global.kafka
-
-class KafkaTopics {
-    companion object {
-        const val CREATE_STATUS = "create-status"
-    }
-}
