+++ application-infrastructure/src/main/kotlin/hs/kr/equus/application/global/kafka/config/KafkaTopics.kt
@@ -1,0 +1,7 @@
+package hs.kr.equus.application.global.kafka.config
+
+class KafkaTopics {
+    companion object {
+        const val UPDATE_STATUS = "update-status"
+    }
+}
